% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{addPhiHist}
\alias{addPhiHist}
\title{Display a histogram of mutation phi and signature exposure for a signature trajectory}
\usage{
addPhiHist(trajectory, trajPlot, truncateStrategy = c("exclude", "stack"))
}
\arguments{
\item{trajectory}{a list containing named elements "mixtures", "changepoints",
and "binData". See @seealso \link{TrackSig}.}

\item{trajPlot}{a ggplot object generated using plotTrajectory @seealso [plotTrajectory()]}

\item{truncateStrategy}{strategy to handle a truncated axis (for example, for a trajectory plotted with anmac = F). \cr
One of "exclude", in which mutations lying off the axis are not displayed, or "stack, in which mutaitons lying off the axis are added to the first histogram bin. \cr
It is advised to use "exclude", unless there are very few clonal mutations.}
}
\value{
tableGrob
}
\description{
Display a histogram of mutation phi and signature exposure for a signature trajectory
}
