% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTrajectories.R
\name{generateContext}
\alias{generateContext}
\title{\code{generateContext} Generate a trinucleotide context from an alphabet. Note: this involves finding all three-member
permutations of the alphabet, which can be inconveinent for large alphabets. Nucleotides are assumed to be provided as complementary pairs,
where the first of each pair is used as the reference to build the context.}
\usage{
generateContext(alphabet)
}
\arguments{
\item{alphabet}{list of pairs of characters to create combinations of as a mutation context type}
}
\value{
data.frame containing all the possible trinucleotide contextes for a mutation in the supplied alphabet
}
\description{
\code{generateContext} Generate a trinucleotide context from an alphabet. Note: this involves finding all three-member
permutations of the alphabet, which can be inconveinent for large alphabets. Nucleotides are assumed to be provided as complementary pairs,
where the first of each pair is used as the reference to build the context.
}
\examples{
context <- TrackSig:::generateContext(c("CG", "TA"))
dim(context) == c(96, 3)
head(context)

}
