% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{\code{plotTrajectory}}
\usage{
plotTrajectory(mixtures, phis = NULL, changepoints = NULL,
  linearX = T, anmac = T, ...)
}
\arguments{
\item{mixtures}{mixtures of mixtures output by TrackSig}

\item{phis}{list of mean phis corresponding to bins in mixtures matrix}

\item{changepoints}{list of changepoints to mark on the trajectory plot}

\item{linearX}{logical whether to plot with a linearly spaced x-axis grid, or with ccf values}

\item{anmac}{logical whether to plot x-axis restricted to ccf space, or use estimated average number of mutant alleles per cell (anmac)}
}
\value{
ggplot object
}
\description{
\code{plotTrajectory}
}
