% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Plot the exolutionary trajectory of a tumour}
\usage{
plotTrajectory(trajectory, linearX = F, anmac = F, show = T)
}
\arguments{
\item{trajectory}{a list containing named elements "mixtures", "changepoints",
and "binData". See @seealso \link{TrackSig}.}

\item{linearX}{logical whether to plot with a linearly spaced x-axis grid, or
with binned phi values}

\item{anmac}{logical whether to plot x-axis restricted to ccf space, or use
estimated average number of mutant alleles per cell (anmac)}

\item{show}{logical whether to print the plot}
}
\value{
ggplot object
}
\description{
For each bin in a set of signature mixtures, the mixture is plotted accross
pseudo-time. Provided changepoints will be highlighted.
}
