% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSig.R
\name{TrackSig}
\alias{TrackSig}
\title{\code{TrackSig} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.}
\usage{
TrackSig(vcfFile, activeInSample, cnaFile = NULL, purity = NULL,
  sampleID = NULL, referenceSignatures = alex,
  scoreMethod = "SigFreq", binSize = 100, desiredMinSegLen = NULL,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{vcfFile}{path to variant calling format (vcf) file}

\item{cnaFile}{path to copy number abberation (cna) file}

\item{sampleID}{name to call sample. If none provided, name will be automatically drawn from the provided vcf file name.}

\item{purityFile}{path to sample purity file}

\item{saveIntermediate}{boolean whether to save intermediate results (mutation types)


activeInSample is list used to subset referenceSignatures}
}
\description{
\code{TrackSig} Take an input vcf file and annotation and generate the counts data.
Create all plotting output that compute_signatures_for_all_examples does.
}
